feat(signer): optionally receive transaction receipt
