+++ core/application/src/env.rs
@@ -197,6 +197,7 @@ impl ApplicationEnv {
             }
         }
 
+        println!("APP RUN");
         Ok(response)
     }
