+++ core/signer/Cargo.toml
@@ -14,6 +14,7 @@ serde.workspace = true
 thiserror.workspace = true
 tokio.workspace = true
 tracing.workspace = true
+quick_cache = "0.6"
 
 [dev-dependencies]
 fleek-crypto.workspace = true
