+++ core/types/src/lib.rs
@@ -20,6 +20,7 @@ mod pool;
 mod reputation;
 mod response;
 mod rpc;
+mod signer;
 mod state;
 mod state_proof;
 mod transaction;
@@ -45,6 +46,7 @@ pub use pool::*;
 pub use reputation::*;
 pub use response::*;
 pub use rpc::*;
+pub use signer::*;
 pub use state::*;
 pub use state_proof::*;
 pub use transaction::*;
